change from int to double

+++ src/main/java/org/wwu/bpm/gta/creditscore/CheckOutstandingCredits.java
@@ -9,7 +9,7 @@ public class CheckOutstandingCredits extends ServletProcessApplication implement
 	@Override
 	public void execute(DelegateExecution execution) throws Exception {
 		// generates random variable in value range from 0 to 10 that represents an applicants outstanding dept situation
-				int OutstandingDebtScore = (int)(Math.random() * 1);	
+				double OutstandingDebtScore = (double)(Math.random() * 1);	
 				execution.setVariable("OutstandingDebtScore", OutstandingDebtScore);
 		
 	}
